usuarios abc

+++ src/main/java/dw/elh/repository/UsuarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import dw.elh.model.Usuario;
 
-public interface UsuarioRepository extends CrudRepository<Usuario, Integer>{
+public interface UsuarioRepository extends CrudRepository<Usuario, Long>{
 	Optional<Usuario> getUsuarioByUsuario(String usuario);
 }
